refactor: 일급 컬렉션 클래스명 변경, 메인 클래스와 Racing 클래스 분리
